play proper scene depending on wedding settings and add some popup message

+++ src/world/Event/Director.cpp
@@ -198,12 +198,12 @@ void Sapphire::Event::Director::setDirectorSequence( uint8_t value )
   m_sequence = value;
 }
 
-void Sapphire::Event::Director::setCustomVar( uint32_t varId, uint32_t value )
+void Sapphire::Event::Director::setCustomVar( uint32_t varId, uint64_t value )
 {
   m_customVarMap[ varId ] = value;
 }
 
-uint32_t Sapphire::Event::Director::getCustomVar( uint32_t varId )
+uint64_t Sapphire::Event::Director::getCustomVar( uint32_t varId )
 {
   auto it = m_customVarMap.find( varId );
   if( it != m_customVarMap.end() )
